+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Changes 0.6.7
+- fixed incorrect docker.cstatus/vstatus/istatus metric ([#74](https://github.com/monitoringartist/zabbix-docker-monitoring/issues/74))
+- removed support for systemd - code has been moved to https://github.com/cavaliercoder/zabbix-module-systemd
+- added timeout for Docker socket query ([#73](https://github.com/monitoringartist/zabbix-docker-monitoring/issues/73))
+- new item key docker.modver
+
 # Changes 0.6.6
 - fixed incorrect CPU multiplier in the template ([#30](https://github.com/monitoringartist/zabbix-docker-monitoring/issues/30)) - update your container CPU triggers
 - added total CPU utilization metric (current sum of user+system ticks)